+++ src/hexen/p_spec.cpp
@@ -1036,16 +1036,13 @@ void P_UpdateSpecials()
                 switch (buttonlist[i].where)
                 {
                     case SWTCH_TOP:
-                        sides[buttonlist[i].line->sidenum[0]].toptexture =
-                            buttonlist[i].btexture;
+                        sides[buttonlist[i].line->sidenum[0]].toptexture = static_cast<short>(buttonlist[i].btexture);
                         break;
                     case SWTCH_MIDDLE:
-                        sides[buttonlist[i].line->sidenum[0]].midtexture =
-                            buttonlist[i].btexture;
+                        sides[buttonlist[i].line->sidenum[0]].midtexture = static_cast<short>(buttonlist[i].btexture);
                         break;
                     case SWTCH_BOTTOM:
-                        sides[buttonlist[i].line->sidenum[0]].bottomtexture =
-                            buttonlist[i].btexture;
+                        sides[buttonlist[i].line->sidenum[0]].bottomtexture = static_cast<short>(buttonlist[i].btexture);
                         break;
                 }
                 //S_StartSound((mobj_t *)&buttonlist[i].soundorg, sfx_switch);
