+++ src/strife/r_data.cpp
@@ -336,8 +336,8 @@ void R_GenerateLookup (int texnum)
 	for ( ; x<x2 ; x++)
 	{
 	    patchcount[x]++;
-	    collump[x] = patch->patch;
-	    colofs[x] = LONG(realpatch->columnofs[x-x1])+3;
+	    collump[x] = static_cast<short>(patch->patch);
+	    colofs[x] = static_cast<unsigned short>(LONG(realpatch->columnofs[x - x1]) + 3);
 	}
     }
 	
@@ -355,7 +355,7 @@ void R_GenerateLookup (int texnum)
 	{
 	    // Use the cached block.
 	    collump[x] = -1;	
-	    colofs[x] = texturecompositesize[texnum];
+	    colofs[x] = static_cast<unsigned short>(texturecompositesize[texnum]);
 	    
 	    if (texturecompositesize[texnum] > 0x10000-texture->height)
 	    {
