+++ src/strife/r_plane.cpp
@@ -182,7 +182,7 @@ void R_ClearPlanes ()
     // opening / clipping determination
     for (i=0 ; i<viewwidth ; i++)
     {
-	floorclip[i] = viewheight;
+	floorclip[i] = static_cast<short>(viewheight);
 	ceilingclip[i] = -1;
     }
